fix invite members bug
