+++ web/components/shared/layout/orgDropdown.tsx
@@ -17,11 +17,11 @@ import CreateOrgForm, {
   ORGANIZATION_ICONS,
 } from "../../templates/organizations/createOrgForm";
 import Link from "next/link";
-import InviteMemberButton from "../inviteMembers";
 import ThemedModal from "../themed/themedModal";
 import { useUserSettings } from "../../../services/hooks/userSettings";
 import useNotification from "../notification/useNotification";
 import { useGetOrgMembers } from "../../../services/hooks/organizations";
+import AddMemberModal from "../../templates/organizationId/addMemberModal";
 
 interface OrgDropdownProps {
   userSettings: {
@@ -56,38 +56,6 @@ export default function OrgDropdown(props: OrgDropdownProps) {
   const currentColor = ORGANIZATION_COLORS.find(
     (icon) => icon.name === orgContext?.currentOrg.color
   );
-
-  const onSubmitHandler = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-
-    const email = e.currentTarget.elements.namedItem(
-      "email"
-    ) as HTMLInputElement;
-
-    if (!email || !email.value || !org?.currentOrg) {
-      setNotification("Failed to add member. Please try again.", "error");
-      return;
-    }
-
-    fetch(
-      `/api/organization/${org?.currentOrg.id}/add_member?email=${email.value}`
-    )
-      .then((res) => res.json())
-      .then((res) => {
-        if (res.error) {
-          if (res.error.length < 30) {
-            setNotification(res.error, "error");
-            console.error(res);
-          } else {
-            setNotification("Error adding member", "error");
-            console.error(res);
-          }
-        } else {
-          setNotification("Member successfully invited.", "success");
-          setAddOpen(false);
-        }
-      });
-  };
 
   return (
     <>
@@ -277,54 +245,11 @@ export default function OrgDropdown(props: OrgDropdownProps) {
           <CreateOrgForm onCancelHandler={setCreateOpen} />
         </div>
       </ThemedModal>
-      <ThemedModal open={addOpen} setOpen={setAddOpen}>
-        <form
-          action="#"
-          method="POST"
-          onSubmit={onSubmitHandler}
-          className="flex flex-col gap-4 w-full"
-        >
-          <p className="font-semibold text-lg">Add New Member</p>
-          <div className="space-y-1.5 text-sm w-[400px]">
-            <label htmlFor="email">User Email</label>
-            <input
-              type="email"
-              name="email"
-              id="email"
-              className={clsx(
-                "block w-full rounded-md border border-gray-300 shadow-sm p-2 text-sm"
-              )}
-              placeholder={"Enter user email"}
-            />
-          </div>
-          <div className="flex justify-end gap-2 pt-4">
-            <button
-              type="button"
-              onClick={() => {
-                setAddOpen(false);
-              }}
-              className="flex flex-row items-center rounded-md bg-white px-4 py-2 text-sm font-semibold border border-gray-300 hover:bg-gray-50 text-gray-900 shadow-sm hover:text-gray-700 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-500"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              onClick={() => {
-                if (!org?.currentOrg) {
-                  setNotification(
-                    "Invalid organization. Please log out and try again",
-                    "error"
-                  );
-                  return;
-                }
-              }}
-              className="items-center rounded-md bg-black px-4 py-2 text-sm flex font-semibold text-white shadow-sm hover:bg-gray-800 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-white"
-            >
-              Add Member
-            </button>
-          </div>
-        </form>
-      </ThemedModal>
+      <AddMemberModal
+        orgId={org?.currentOrg.id || ""}
+        open={addOpen}
+        setOpen={setAddOpen}
+      />
     </>
   );
 }
